The login activity has been scheduled.
