+++ app/src/main/res/drawable/bx_arrow_back.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M21,11H6.414l5.293,-5.293 -1.414,-1.414L2.586,12l7.707,7.707 1.414,-1.414L6.414,13H21z"/>
+</vector>
