+++ app/src/main/res/drawable/bxl_facebook_circle.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="30dp"
+    android:height="30dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#2D61B1"
+        android:pathData="M12.001,2.002c-5.522,0 -9.999,4.477 -9.999,9.999 0,4.99 3.656,9.126 8.437,9.879v-6.988h-2.54v-2.891h2.54V9.798c0,-2.508 1.493,-3.891 3.776,-3.891 1.094,0 2.24,0.195 2.24,0.195v2.459h-1.264c-1.24,0 -1.628,0.772 -1.628,1.563v1.875h2.771l-0.443,2.891h-2.328v6.988C18.344,21.129 22,16.992 22,12.001c0,-5.522 -4.477,-9.999 -9.999,-9.999z"/>
+</vector>
