+++ app/src/main/res/drawable/bxl_instagram.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="30dp"
+    android:height="30dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#2D61B1"
+        android:pathData="M11.999,7.377a4.623,4.623 0,1 0,0 9.248,4.623 4.623,0 0,0 0,-9.248zM11.999,15.004a3.004,3.004 0,1 1,0 -6.008,3.004 3.004,0 0,1 0,6.008z"/>
+    <path
+        android:fillColor="#2D61B1"
+        android:pathData="M16.806,7.207m-1.078,0a1.078,1.078 0,1 1,2.156 0a1.078,1.078 0,1 1,-2.156 0"/>
+    <path
+        android:fillColor="#2D61B1"
+        android:pathData="M20.533,6.111A4.605,4.605 0,0 0,17.9 3.479a6.606,6.606 0,0 0,-2.186 -0.42c-0.963,-0.042 -1.268,-0.054 -3.71,-0.054s-2.755,0 -3.71,0.054a6.554,6.554 0,0 0,-2.184 0.42,4.6 4.6,0 0,0 -2.633,2.632 6.585,6.585 0,0 0,-0.419 2.186c-0.043,0.962 -0.056,1.267 -0.056,3.71 0,2.442 0,2.753 0.056,3.71 0.015,0.748 0.156,1.486 0.419,2.187a4.61,4.61 0,0 0,2.634 2.632,6.584 6.584,0 0,0 2.185,0.45c0.963,0.042 1.268,0.055 3.71,0.055s2.755,0 3.71,-0.055a6.615,6.615 0,0 0,2.186 -0.419,4.613 4.613,0 0,0 2.633,-2.633c0.263,-0.7 0.404,-1.438 0.419,-2.186 0.043,-0.962 0.056,-1.267 0.056,-3.71s0,-2.753 -0.056,-3.71a6.581,6.581 0,0 0,-0.421 -2.217zM19.315,15.643a5.043,5.043 0,0 1,-0.311 1.688,2.987 2.987,0 0,1 -1.712,1.711 4.985,4.985 0,0 1,-1.67 0.311c-0.95,0.044 -1.218,0.055 -3.654,0.055 -2.438,0 -2.687,0 -3.655,-0.055a4.96,4.96 0,0 1,-1.669 -0.311,2.985 2.985,0 0,1 -1.719,-1.711 5.08,5.08 0,0 1,-0.311 -1.669c-0.043,-0.95 -0.053,-1.218 -0.053,-3.654 0,-2.437 0,-2.686 0.053,-3.655a5.038,5.038 0,0 1,0.311 -1.687c0.305,-0.789 0.93,-1.41 1.719,-1.712a5.01,5.01 0,0 1,1.669 -0.311c0.951,-0.043 1.218,-0.055 3.655,-0.055s2.687,0 3.654,0.055a4.96,4.96 0,0 1,1.67 0.311,2.991 2.991,0 0,1 1.712,1.712 5.08,5.08 0,0 1,0.311 1.669c0.043,0.951 0.054,1.218 0.054,3.655 0,2.436 0,2.698 -0.043,3.654h-0.011z"/>
+</vector>
