+++ app/src/main/res/drawable/bxl_twitter.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="30dp"
+    android:height="30dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="#2D61B1"
+        android:pathData="M19.633,7.997c0.013,0.175 0.013,0.349 0.013,0.523 0,5.325 -4.053,11.461 -11.46,11.461 -2.282,0 -4.402,-0.661 -6.186,-1.809 0.324,0.037 0.636,0.05 0.973,0.05a8.07,8.07 0,0 0,5.001 -1.721,4.036 4.036,0 0,1 -3.767,-2.793c0.249,0.037 0.499,0.062 0.761,0.062 0.361,0 0.724,-0.05 1.061,-0.137a4.027,4.027 0,0 1,-3.23 -3.953v-0.05c0.537,0.299 1.16,0.486 1.82,0.511a4.022,4.022 0,0 1,-1.796 -3.354c0,-0.748 0.199,-1.434 0.548,-2.032a11.457,11.457 0,0 0,8.306 4.215c-0.062,-0.3 -0.1,-0.611 -0.1,-0.923a4.026,4.026 0,0 1,4.028 -4.028c1.16,0 2.207,0.486 2.943,1.272a7.957,7.957 0,0 0,2.556 -0.973,4.02 4.02,0 0,1 -1.771,2.22 8.073,8.073 0,0 0,2.319 -0.624,8.645 8.645,0 0,1 -2.019,2.083z"/>
+</vector>
